fixed a circular import and two nonexistent directories
