+++ xrsdkit/models/__init__.py
@@ -44,6 +44,7 @@ regression_models = OrderedDict()
 classification_models = OrderedDict()
 
 # --- LOAD REGRESSION MODELS --- #
+if not os.path.exists(regression_models_dir): os.mkdir(regression_models_dir)
 for sys_cls in os.listdir(regression_models_dir):
     regression_models[sys_cls] = {}
     sys_cls_dir = os.path.join(regression_models_dir,sys_cls)
@@ -80,6 +81,7 @@ for sys_cls in os.listdir(regression_models_dir):
 
 
 # --- LOAD CLASSIFICATION MODELS --- #
+if not os.path.exists(classification_models_dir): os.mkdir(classification_models_dir)
 yml_path = os.path.join(classification_models_dir,'system_classification.yml')
 classification_models['system_classification'] = Classifier('system_classification',yml_path)
 for sys_cls in os.listdir(classification_models_dir):
@@ -451,6 +453,76 @@ def save_classification_models(models=classification_models, test=False):
                     yml_path = os.path.join(sys_dir_path,pop_id,cls_label+'.yml')
                     txt_path = os.path.join(sys_dir_path,pop_id,cls_label+'.txt')
                     save_model_data(m,yml_path,txt_path)
+
+# TODO refactor the modeling dataset index: it can no longer be divided simply by system class
+def save_modeling_datasets(df,grp_cols,all_groups,all_samples,test=True):
+    dir_path = modeling_data_dir
+    if test:
+        dir_path = os.path.join(dir_path,'models','modeling_data','testing_data')
+    file_path = os.path.join(dir_path,'dataset_statistics.txt')
+
+    with open(file_path, 'w') as txt_file:
+        txt_file.write('Downsampling statistics:\n\n')
+        for grpk,samp in zip(all_groups.groups.keys(),all_samples):
+            txt_file.write(grp_cols+'\n')
+            txt_file.write(grpk+'\n')
+            txt_file.write(len(samp)+' / '+len(all_groups.groups[grpk])+'\n')
+
+    modeling_dsid_file = os.path.join(modeling_data_dir,'modeling_dataset_ids.yml')
+    all_dsids = yaml.load(open(modeling_dsid_file,'rb'))
+
+    ds_map_filepath = os.path.join(modeling_data_dir,'dsid_map.yml')
+    ds_map = yaml.load(open(ds_map_filepath,'rb'))
+
+
+    # TODO: Take all_dsids one at a time,
+    # and associate each one with a group.
+    # (NOTE: If the group already exists in the ds_map,
+    # should we re-use that dsid?)
+    # If we run out of available modeling datasets,
+    # we will add more to the list by hand.
+    # ds_map should be an embedded dict,
+    # keyed by all_groups.groups.keys.
+
+    # For each dataset that gets assigned to a group,
+    # set its title to 'xrsdkit modeling dataset',
+    # set its description to list the group labels,
+    # create a new version,
+    # and upload the group.
+
+    # Then, upload the entire sample for system_classifier,
+    # and upload each system_class into a dataset 
+    # for that system's basis_classifiers.
+    #            pif.dump(pp, open(jsf,'w'))
+    #            client.data.upload(ds_id, jsf)
+    #    with open(ds_map_filepath, 'w') as yaml_file:
+    #        yaml.dump(dataset_ids, yaml_file)
+
+
+# TODO: generate all unique sets of labels and the corresponding dataframe groups 
+def group_by_labels(df):
+    grp_cols = ['experiment_id','system_classification']
+    for col in df.columns:
+        if re.compile('pop._basis_classification').match(col): grp_cols.append(col)
+        if re.compile('pop._lattice').match(col): grp_cols.append(col)
+        if re.compile('pop._interaction').match(col): grp_cols.append(col)
+    all_groups = df.groupby(grp_cols)
+    return grp_cols, all_groups
+
+    #all_labels = []
+    #all_groups = []
+    # first, group by experiment_id: 
+    #expt_grps = df.groupby('experiment_id')
+    #for expt_id,expt_grp in expt_grps.items():
+    #    expt_lbls = {'experiment_id':expt_id}
+    #    # next, group by system_classification:
+    #    sys_cls_grps = expt_grp.groupby('system_classification')
+    #    for sys_cls, sys_cls_grp in sys_cls_grps.items():
+    #        sys_cls_lbls = copy.deepcopy(expt_lbls)
+    #        sys_cls_lbls['system_classification'] = sys_cls
+    #        # TODO:
+    #        # finally, group by population classifications:
+    #        # NOTE: this is tricky
 
 # helper function - to set parameters for scalers and models
 def set_param(m_s, param):
